Remove text and do a bit of refactor
