+++ src/components/Extension/Extension.tsx
@@ -23,7 +23,7 @@ import { StripeExtension } from './StripeExtension'
 import { IdentityVerification } from './IdentityVerification'
 import { ShellyExtension } from './ShellyExtension'
 import { RequestingProviderLookUp } from './PrivateExtensions/WP'
-import { StartCareFlowAndSession } from './SharedActions'
+import { WaitForActivityToComplete } from './SharedActions'
 
 interface ExtensionProps {
   activity: Activity
@@ -74,7 +74,9 @@ export const Extension: FC<ExtensionProps> = ({ activity }) => {
         )
       case AnonymousActionKeys.START_CARE_FLOW_AND_SESSION:
         return (
-          <StartCareFlowAndSession activityDetails={extensionActivityDetails} />
+          <WaitForActivityToComplete
+            activityDetails={extensionActivityDetails}
+          />
         )
       // TODO add a specific case for the new plugin action
       default:
