--- src/components/Extension/SharedActions/StartCareFlowAndSession/StartCareFlowAndSession.module.css
@@ -1,3 +1,0 @@
-.redirect {
-  text-align: center;
-}
