--- src/components/Extension/SharedActions/StartCareFlowAndSession/index.tsx
@@ -1,1 +1,0 @@
-export { StartCareFlowAndSession } from './StartCareFlowAndSession'
