--- src/components/Extension/SharedActions/StartCareFlowAndSession/types.ts
@@ -1,4 +1,0 @@
-export {
-  type DataPoints,
-  useCompleteExtensionActivity,
-} from '../../../../hooks/useCompleteExtensionActivity'
