+++ src/components/Extension/SharedActions/WaitForActivityToComplete/index.tsx
@@ -1,2 +1,1 @@
-export { Redirect } from './Redirect'
 export { WaitForActivityToComplete } from './WaitForActivityToComplete'
